@@ -6,18 +6,17 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.1 <4.0.0"
-dependencies:
+  sdk: ">=3.0.0 <4.0.0"
 
-  provider: ^6.0.4
-  share_plus: 6.3.4
-  carousel_slider: ^4.2.1
-  photo_view: ^0.14.0
+dependencies:
   flutter:
     sdk: flutter
-  http: ^0.13.5
-  flutter_html: ^3.0.0-alpha.5
-  # flutter_markdown: ^0.6.14
+  provider: ^6.1.1
+  share_plus: ^7.2.1
+  carousel_slider: ^4.2.1
+  photo_view: ^0.14.0
+  http: ^1.1.0
+  flutter_html: ^3.0.0-beta.2
   intl: ^0.18.0
   email_validator: ^2.1.17
   flutter_riverpod: ^2.4.0
@@ -28,14 +27,16 @@ dependencies:
   flutter_vector_icons: ^2.0.0
   page_transition: ^2.0.5
   html2md: ^1.3.2
-  shared_preferences: ^2.3.3
+  flutter_secure_storage: ^8.0.0
+  # shared_preferences: ^2.3.3
+  # shared_preferences_web: ^2.2.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   freezed: ^2.4.1
   build_runner: ^2.4.6
-  flutter_lints: ^1.0.0
+  flutter_lints: ^3.0.1
 
 flutter:
   uses-material-design: true
