@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:frontend/constants.dart';
 
 class ScanPage extends StatefulWidget {
-  const ScanPage({Key? key}) : super(key: key);
+  const ScanPage({super.key});
 
   @override
   State<ScanPage> createState() => _ScanPageState();
