@@ -7,7 +7,7 @@ import 'package:frontend/Services/api_service.dart';
 import 'package:email_validator/email_validator.dart';
 
 class ForgotPassword extends StatefulWidget {
-  const ForgotPassword({Key? key}) : super(key: key);
+  const ForgotPassword({super.key});
 
   @override
   State<ForgotPassword> createState() => _ForgotPasswordState();
