@@ -6,23 +6,17 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up()
     {
         Schema::table('comments', function (Blueprint $table) {
-            $table->unsignedBigInteger('product_id')->after('id'); // Thêm cột product_id
-            $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade'); // Liên kết với bảng products
+            $table->string('email')->nullable()->change();
         });
     }
-    
+
     public function down()
     {
         Schema::table('comments', function (Blueprint $table) {
-            $table->dropForeign(['product_id']);
-            $table->dropColumn('product_id');
+            $table->string('email')->nullable(false)->change();
         });
     }
-    
 };
