@@ -20,18 +20,24 @@
                             <i data-lucide="x" class="w-4 h-4"></i> 
                         </button> 
                     </div>
-                    
-                    @endif
+                @endif
                 
                 <div class="hidden xl:flex flex-col min-h-screen">
                     <a href="" class="-intro-x flex items-center pt-5">
-                        <span class="text-white text-lg ml-3">{{$detail->company_name}} </span> 
+                        @if ($detail)
+                            <span class="text-white text-lg ml-3">{{$detail->company_name}}</span>
+                        @else
+                            <span class="text-white text-lg ml-3">Default Company Name</span>
+                        @endif
                     </a>
                     <div class="my-auto">
-                        <img alt="Midone - HTML Admin Template" class="-intro-x w-1/2 -mt-16" src="{{$detail->logo}}">
+                        @if ($detail && $detail->logo)
+                            <img alt="Logo" class="-intro-x w-1/2 -mt-16" src="{{$detail->logo}}">
+                        @else
+                            <img alt="Logo" class="-intro-x w-1/2 -mt-16" src="default-logo.png"> <!-- Replace with a default logo -->
+                        @endif
                         <div class="-intro-x text-white font-medium text-4xl leading-tight mt-10">
                             Quản lý bán hàng <br/>
-                            
                         </div>
                         <div class="-intro-x mt-5 text-lg text-white text-opacity-70 dark:text-slate-400"> </div>
                     </div>
@@ -103,3 +109,4 @@
         @include('backend.layouts.footer')
     </body>
 </html>
+c
