@@ -3,14 +3,16 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../providers/loginProvider.dart';
 
 class MyProfile extends ConsumerWidget {
+  const MyProfile({super.key});
+
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final user = ref.watch(userProvider);
 
     if (user == null) {
       return Scaffold(
-        appBar: AppBar(title: Text("Hồ sơ của tôi")),
-        body: Center(
+        appBar: AppBar(title: const Text("Hồ sơ của tôi")),
+        body: const Center(
           child: Text(
             "Không có dữ liệu người dùng.",
             style: TextStyle(fontSize: 18, color: Colors.grey),
@@ -21,7 +23,7 @@ class MyProfile extends ConsumerWidget {
 
     return Scaffold(
       appBar: AppBar(
-        title: Text("Hồ sơ của tôi"),
+        title: const Text("Hồ sơ của tôi"),
         backgroundColor: Colors.green, // Chọn màu appBar đẹp
       ),
       body: Padding(
@@ -38,34 +40,34 @@ class MyProfile extends ConsumerWidget {
                         height: 100,
                         fit: BoxFit.cover,
                       )
-                    : Icon(Icons.person, size: 100, color: Colors.grey),
+                    : const Icon(Icons.person, size: 100, color: Colors.grey),
               ),
             ),
-            SizedBox(height: 20),
+            const SizedBox(height: 20),
 
             // Họ và tên
             _buildProfileField("Họ và tên", user.full_name),
-            SizedBox(height: 10),
+            const SizedBox(height: 10),
 
             // Username
             _buildProfileField("Username", user.username),
-            SizedBox(height: 10),
+            const SizedBox(height: 10),
 
             // Email
             _buildProfileField("Email", user.email),
-            SizedBox(height: 10),
+            const SizedBox(height: 10),
 
             // Điện thoại
             _buildProfileField("Điện thoại", user.phone ?? "Chưa có thông tin"),
-            SizedBox(height: 10),
+            const SizedBox(height: 10),
 
             // Vai trò
             _buildProfileField("Vai trò", user.role),
-            SizedBox(height: 10),
+            const SizedBox(height: 10),
 
             // Trạng thái
             _buildProfileField("Trạng thái", user.status ?? "Đang hoạt động"),
-            SizedBox(height: 20),
+            const SizedBox(height: 20),
 
             // Nút chỉnh sửa (Nếu cần thiết)
             ElevatedButton(
@@ -73,12 +75,12 @@ class MyProfile extends ConsumerWidget {
                 // Chức năng chỉnh sửa nếu cần
                 // Ví dụ: mở trang chỉnh sửa hồ sơ
               },
-              child: Text("Chỉnh sửa hồ sơ"),
               style: ElevatedButton.styleFrom(
                 backgroundColor: Colors.green, // Màu nút
-                padding: EdgeInsets.symmetric(vertical: 12, horizontal: 40),
-                textStyle: TextStyle(fontSize: 16),
+                padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 40),
+                textStyle: const TextStyle(fontSize: 16),
               ),
+              child: Text("Chỉnh sửa hồ sơ"),
             ),
           ],
         ),
@@ -90,7 +92,7 @@ class MyProfile extends ConsumerWidget {
   Widget _buildProfileField(String label, String value) {
     return Card(
       elevation: 4,
-      margin: EdgeInsets.symmetric(vertical: 5),
+      margin: const EdgeInsets.symmetric(vertical: 5),
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(12),
       ),
@@ -98,24 +100,24 @@ class MyProfile extends ConsumerWidget {
         padding: const EdgeInsets.all(12.0),
         child: Row(
           children: [
-            Icon(Icons.info_outline, color: Colors.green), // Icon cho trường
-            SizedBox(width: 10),
+            const Icon(Icons.info_outline, color: Colors.green), // Icon cho trường
+            const SizedBox(width: 10),
             Expanded(
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
                   Text(
                     label,
-                    style: TextStyle(
+                    style: const TextStyle(
                       fontSize: 16,
                       fontWeight: FontWeight.w600,
                       color: Colors.teal,
                     ),
                   ),
-                  SizedBox(height: 4),
+                  const SizedBox(height: 4),
                   Text(
                     value,
-                    style: TextStyle(fontSize: 16, color: Colors.black87),
+                    style: const TextStyle(fontSize: 16, color: Colors.black87),
                     overflow: TextOverflow.ellipsis,
                   ),
                 ],
