@@ -12,7 +12,7 @@ import 'package:frontend/main.dart';
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(const MyApp(token: '',));
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
